全新的加载逻辑

+++ AutumnBox.GUI/Util/AppUnloader.cs
@@ -4,6 +4,7 @@
 ** desc： ...
 *************************************************/
 using AutumnBox.Basic.ManagedAdb;
+using AutumnBox.GUI.Properties;
 using AutumnBox.GUI.Util.OS;
 using AutumnBox.OpenFramework.Management;
 
@@ -11,13 +12,37 @@ namespace AutumnBox.GUI.Util
 {
     public class AppUnloader
     {
+        public static AppUnloader Instance
+        {
+            get
+            {
+                if (_instance == null)
+                {
+                    _instance = new AppUnloader();
+                }
+                return _instance;
+            }
+        }
+        private static AppUnloader _instance;
+        private AppUnloader() { }
         public void Unload()
         {
             try
             {
-                OpenFx.Unload();
-                Adb.Server.Kill();
-                TaskKill.Kill("adb.exe");
+                if (Settings.Default.IsFirstLaunch && Settings.Default.GuidePassed)
+                {
+                    Settings.Default.IsFirstLaunch = false;
+                }
+                Settings.Default.Save();
+                if (Bus.OpenFxObserver.Instance.IsLoaded)
+                {
+                    OpenFx.Unload();
+                }
+                if (Adb.Server.IsEnable)
+                {
+                    Adb.Server.Kill();
+                    TaskKill.Kill("adb.exe");
+                }
             }
             catch { }
         }
